reset fixtures
