+++ psiresp/tests/test_job.py
@@ -18,7 +18,7 @@ from psiresp.tests.datafiles import (AMM_NME_OPT_ESPA1_CHARGES,
                                      AMM_NME_OPT_RESPA1_CHARGES,
                                      MANUAL_JOBS_WKDIR,
                                      )
-from psiresp.server import fractal_client, empty_client
+# from psiresp.server import fractal_client, empty_client
 
 
 class TestSingleResp:
